requirements.txt
duckdb
pandas
requests
python-dotenv
prefect
dbt-core
dbt-duckdb
streamlit
plotly
